src/crypto/crypto_manager.py
"""
Crypto Manager module for handling cryptographic operations and privacy-preserving techniques.
"""

import logging
import os
import json
import hashlib
from Crypto.Cipher import AES
from Crypto.Util.Padding import pad, unpad
from Crypto.Random import get_random_bytes
import base64

# Note: In a production environment, these would be replaced with actual libraries
# for zero-knowledge proofs and homomorphic encryption
# This is a simplified implementation for demonstration purposes

logger = logging.getLogger(__name__)

class CryptoManager:
    """Manages cryptographic operations including zero-knowledge proofs and homomorphic encryption."""
    
    def __init__(self, config):
        """
        Initialize the CryptoManager with the provided configuration.
        
        Args:
            config (dict): Cryptography configuration including key paths and security parameters
        """
        self.config = config
        self._initialize_keys()
        logger.info("CryptoManager initialized")
    
    def _initialize_keys(self):
        """Initialize cryptographic keys."""
        try:
            # Create key directories if they don't exist
            os.makedirs(os.path.dirname(self.config['zkp']['proving_key_path']), exist_ok=True)
            os.makedirs(os.path.dirname(self.config['homomorphic']['key_path']), exist_ok=True)
            
            # In a real implementation, this would load actual keys
            # For this demo, we'll generate placeholder keys
            self.aes_key = get_random_bytes(32)  # 256-bit key for AES
            
            logger.info("Cryptographic keys initialized")
        except Exception as e:
            logger.error(f"Error initializing cryptographic keys: {e}")
    
    def hash_identity_data(self, identity_data):
        """
        Create a secure hash of identity data.
        
        Args:
            identity_data (dict): The identity data to hash
            
        Returns:
            str: The hash of the identity data
        """
        try:
            # Convert the identity data to a consistent JSON string
            normalized_data = json.dumps(identity_data, sort_keys=True)
            
            # Create SHA-256 hash
            hash_obj = hashlib.sha256(normalized_data.encode())
            hash_hex = hash_obj.hexdigest()
            
            return hash_hex
        except Exception as e:
            logger.error(f"Error hashing identity data: {e}")
            return None
    
    def encrypt_data(self, data):
        """
        Encrypt data using AES-256.
        
        Args:
            data (str or dict): The data to encrypt
            
        Returns:
            str: Base64-encoded encrypted data
        """
        try:
            # Convert dict to JSON string if necessary
            if isinstance(data, dict):
                plaintext = json.dumps(data).encode()
            else:
                plaintext = data.encode()
            
            # Generate a random IV
            iv = get_random_bytes(16)
            
            # Create cipher and encrypt
            cipher = AES.new(self.aes_key, AES.MODE_CBC, iv)
            ciphertext = cipher.encrypt(pad(plaintext, AES.block_size))
            
            # Combine IV and ciphertext and encode as base64
            encrypted_data = base64.b64encode(iv + ciphertext).decode()
            
            return encrypted_data
        except Exception as e:
            logger.error(f"Error encrypting data: {e}")
            return None
    
    def decrypt_data(self, encrypted_data):
        """
        Decrypt AES-256 encrypted data.
        
        Args:
            encrypted_data (str): Base64-encoded encrypted data
            
        Returns:
            str or dict: The decrypted data
        """
        try:
            # Decode from base64
            data = base64.b64decode(encrypted_data)
            
            # Extract IV (first 16 bytes) and ciphertext
            iv = data[:16]
            ciphertext = data[16:]
            
            # Create cipher and decrypt
            cipher = AES.new(self.aes_key, AES.MODE_CBC, iv)
            plaintext = unpad(cipher.decrypt(ciphertext), AES.block_size).decode()
            
            # Try to parse as JSON
            try:
                return json.loads(plaintext)
            except:
                return plaintext
        except Exception as e:
            logger.error(f"Error decrypting data: {e}")
            return None
    
    def create_zero_knowledge_proof(self, statement, private_witness):
        """
        Create a zero-knowledge proof that a statement is true without revealing the witness.
        
        Args:
            statement (str): The public statement to prove
            private_witness (str): The private information that proves the statement
            
        Returns:
            dict: The zero-knowledge proof
        """
        try:
            # In a real implementation, this would use actual ZKP libraries
            # For demonstration purposes, we'll create a simplified simulation
            
            # Hash the witness
            witness_hash = hashlib.sha256(private_witness.encode()).hexdigest()
            
            # Create a "proof" (this is NOT a real ZKP, just a demonstration)
            # A real ZKP would use libraries like ZoKrates, Snarkjs, etc.
            proof = {
                "statement": statement,
                "proof_data": hashlib.sha256((witness_hash + statement).encode()).hexdigest(),
                "verification_hint": witness_hash[:8]  # Leaking a hint (NOT secure in real ZKP)
            }
            
            return proof
        except Exception as e:
            logger.error(f"Error creating zero-knowledge proof: {e}")
            return None
    
    def verify_zero_knowledge_proof(self, proof, verification_context=None):
        """
        Verify a zero-knowledge proof.
        
        Args:
            proof (dict): The zero-knowledge proof to verify
            verification_context (dict, optional): Additional context for verification
            
        Returns:
            bool: Whether the proof is valid
        """
        try:
            # In a real implementation, this would use actual ZKP libraries
            # For demonstration purposes, we'll accept our simplified simulation
            
            # A real ZKP verification would validate the proof cryptographically
            # without needing the original witness
            
            # For our demo, we'll simply check that the proof structure is valid
            required_keys = ["statement", "proof_data", "verification_hint"]
            if not all(key in proof for key in required_keys):
                logger.warning("Invalid proof structure")
                return False
                
            # Simulating successful verification
            return True
        except Exception as e:
            logger.error(f"Error verifying zero-knowledge proof: {e}")
            return False
    
    def homomorphic_encrypt(self, data):
        """
        Encrypt data using simulated homomorphic encryption.
        
        Args:
            data (int or float): Numerical data to encrypt
            
        Returns:
            dict: Homomorphically encrypted data
        """
        try:
            # In a real implementation, this would use actual homomorphic encryption libraries
            # like TenSEAL, SEAL, etc.
            # For demonstration purposes, we'll create a simplified simulation
            
            # Simulated homomorphic encryption (NOT secure)
            # A real implementation would use proper HE libraries
            noise = sum([ord(c) for c in get_random_bytes(4).hex()])
            encrypted_value = data * 1000 + noise
            
            return {
                "encrypted_value": encrypted_value,
                "metadata": {
                    "type": "simulated_he",
                    "scale": 1000
                }
            }
        except Exception as e:
            logger.error(f"Error in homomorphic encryption: {e}")
            return None
    
    def homomorphic_operation(self, operation, encrypted_values):
        """
        Perform operations on homomorphically encrypted data.
        
        Args:
            operation (str): The operation to perform ('add', 'multiply', etc.)
            encrypted_values (list): List of encrypted values to operate on
            
        Returns:
            dict: Result of the homomorphic operation
        """
        try:
            # In a real implementation, this would use actual homomorphic encryption libraries
            # For demonstration purposes, we'll create a simplified simulation
            
            # Extract encrypted values
            values = [ev["encrypted_value"] for ev in encrypted_values]
            
            # Perform the operation (simulated)
            if operation == "add":
                result = sum(values)
            elif operation == "multiply":
                result = 1
                for v in values:
                    result *= v
            else:
                raise ValueError(f"Unsupported homomorphic operation: {operation}")
            
            return {
                "encrypted_value": result,
                "metadata": {
                    "type": "simulated_he",
                    "scale": 1000,
                    "operation": operation
                }
            }
        except Exception as e:
            logger.error(f"Error in homomorphic operation: {e}")
            return None
    
    def homomorphic_decrypt(self, encrypted_data):
        """
        Decrypt homomorphically encrypted data.
        
        Args:
            encrypted_data (dict): The encrypted data
            
        Returns:
            float: The decrypted value
        """
        try:
            # In a real implementation, this would use actual homomorphic encryption libraries
            # For demonstration purposes, we'll use our simplified simulation
            
            # Extract the encrypted value and scale
            encrypted_value = encrypted_data["encrypted_value"]
            scale = encrypted_data["metadata"]["scale"]
            
            # Simulated decryption (only works with our simplified encryption)
            # Real HE would use proper decryption methods
            approximate_value = encrypted_value / scale
            
            return approximate_value
        except Exception as e:
            logger.error(f"Error in homomorphic decryption: {e}")
            return None 