requirements.txt
# Core dependencies
pycryptodome>=3.15.0       # For encryption and cryptographic operations
web3>=5.30.0               # For blockchain interactions
eth-brownie>=1.19.0        # Smart contract development framework
fastapi>=0.88.0            # API framework
uvicorn>=0.20.0            # ASGI server
python-multipart>=0.0.5    # For handling form data
pydantic>=1.10.2           # Data validation and settings management

# AI and ML
tensorflow>=2.10.0          # Machine learning framework
scikit-learn>=1.2.0         # Machine learning utilities
opencv-python>=4.6.0.66     # Computer vision for facial recognition
face-recognition>=1.3.0     # Facial recognition library
anomaly-detection>=0.1.1    # Anomaly detection for fraud prevention

# Privacy-preserving
#pysnark>=0.4.0              # Zero-knowledge proofs
tenseal>=0.3.10             # Homomorphic encryption

# Blockchain
py-solc-x>=1.1.1            # Solidity compiler
python-ethereumbook>=0.1.0  # Ethereum utilities

# Testing
pytest>=7.2.0               # Testing framework
pytest-cov>=4.0.0           # Coverage reporting

# Documentation
sphinx>=5.3.0               # Documentation generator   